requirements.txt
"""
RetinaXpert Python Dependencies
Install with: pip install -r requirements.txt
"""

# Core Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# Machine Learning
scikit-learn>=1.0.0
numpy>=1.20.0
pandas>=1.3.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Utilities
tqdm>=4.60.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Additional (Optional)
tensorflow-addons>=0.15.0  # For F1Score metric
jupyter>=1.0.0  # For Jupyter notebooks
